+++ integration-mc1-14/src/main/kotlin/org/eln2/mc114/Eln2Block.kt
@@ -1,0 +1,11 @@
+package org.eln2.mc114
+
+import org.eln2.mc114.singleBlock.FlubberBlock
+
+class Eln2Block {
+	companion object {
+		val flubberBlock = FlubberBlock()
+
+		fun register() {}
+	}
+}
