+++ integration-mc1-14/src/main/kotlin/org/eln2/mc114/Eln2Items.kt
@@ -1,0 +1,11 @@
+package org.eln2.mc114
+
+import org.eln2.mc114.singleBlock.FlubberItem
+
+class Eln2Items {
+	companion object {
+		val flubberItem = FlubberItem()
+
+		fun register() {}
+	}
+}
